move pkg/master/server to cmd/kube-apiserver/app
